refactor(auth): Use credential store for mfa management
