+++ aegis/init.go
@@ -37,15 +37,15 @@ import (
 )
 
 // Initialize 初始化 Auth 模块，返回 Handler
-func Initialize(hermesSvc contract.HermesProvider, userSvc contract.UserProvider, credentialSvc contract.CredentialService, cacheManager *cache.Manager) (*Handler, error) {
+func Initialize(hermesSvc contract.HermesProvider, userSvc contract.UserProvider, credentialStore contract.CredentialStore, cacheManager *cache.Manager) (*Handler, error) {
 	if hermesSvc == nil {
 		return nil, fmt.Errorf("hermes service is required")
 	}
 	if userSvc == nil {
 		return nil, fmt.Errorf("user service is required")
 	}
-	if credentialSvc == nil {
-		return nil, fmt.Errorf("credential service is required")
+	if credentialStore == nil {
+		return nil, fmt.Errorf("credential store is required")
 	}
 	if cacheManager == nil {
 		return nil, fmt.Errorf("cache manager is required")
@@ -60,7 +60,7 @@ func Initialize(hermesSvc contract.HermesProvider, userSvc contract.UserProvider
 		logger.Info("[Auth] 邮件发送器初始化完成")
 	}
 
-	webauthnSvc, captchaVerifier, totpVerifier, err := initProviders(credentialSvc, cacheManager, userSvc)
+	webauthnSvc, captchaVerifier, err := initProviders(cacheManager, userSvc)
 	if err != nil {
 		return nil, err
 	}
@@ -68,6 +68,10 @@ func Initialize(hermesSvc contract.HermesProvider, userSvc contract.UserProvider
 	throttler := throttle.NewThrottler(cacheManager.Redis())
 	ac := accessctl.NewManager(throttler)
 	logger.Info("[Auth] 访问控制管理器初始化完成")
+
+	mfaSvc := NewMFAService(credentialStore, cacheManager, webauthnSvc)
+	totpVerifier := totp.NewVerifier(mfaSvc)
+	logger.Info("[Auth] TOTP 验证器初始化完成")
 
 	registry := initRegistry(hermesSvc, userSvc, cacheManager, emailSender, webauthnSvc, captchaVerifier, totpVerifier, ac, tokenSvc)
 
@@ -80,7 +84,6 @@ func Initialize(hermesSvc contract.HermesProvider, userSvc contract.UserProvider
 	authenticateSvc := authenticate.NewService(cacheManager, ac)
 	authorizeSvc := authorize.NewService(cacheManager, hermesSvc, userService, tokenSvc, pool, 5*time.Minute)
 	challengeSvc := challenge.NewService(cacheManager, registry)
-	mfaSvc := NewMFAService(credentialSvc, webauthnSvc)
 	profileHandler := NewProfileHandler(userSvc, mfaSvc)
 
 	handler := NewHandler(authenticateSvc, authorizeSvc, challengeSvc, userService, cacheManager, tokenSvc, mfaSvc, profileHandler, pool)
@@ -153,11 +156,11 @@ func initKeyProviders(cm *cache.Manager) (key.MultiOf, key.MultiOf, key.MultiOf,
 	return domainSign, domainVerify, serviceKey, appVerify
 }
 
-// initProviders 初始化底层认证能力（WebAuthn、Captcha、TOTP）
-func initProviders(credentialSvc contract.CredentialService, cacheManager *cache.Manager, userSvc contract.UserProvider) (*webauthn.Service, captcha.Verifier, factor.TOTPVerifier, error) {
+// initProviders 初始化底层认证能力（WebAuthn、Captcha）
+func initProviders(cacheManager *cache.Manager, userSvc contract.UserProvider) (*webauthn.Service, captcha.Verifier, error) {
 	webauthnSvc, err := webauthn.NewService(cacheManager, userSvc)
 	if err != nil {
-		return nil, nil, nil, fmt.Errorf("init webauthn service: %w", err)
+		return nil, nil, fmt.Errorf("init webauthn service: %w", err)
 	}
 	logger.Info("[Auth] WebAuthn 初始化完成")
 
@@ -167,11 +170,7 @@ func initProviders(credentialSvc contract.CredentialService, cacheManager *cache
 		logger.Infof("[Auth] Captcha 验证器初始化完成: provider=%s", captchaVerifier.GetProvider())
 	}
 
-	// TOTP
-	totpVerifier := totp.NewVerifier(credentialSvc)
-	logger.Info("[Auth] TOTP 验证器初始化完成")
-
-	return webauthnSvc, captchaVerifier, totpVerifier, nil
+	return webauthnSvc, captchaVerifier, nil
 }
 
 // initRegistry 初始化全局 Registry（注册胶水层 Authenticator）
